Move Chromatic to deprecated (#789)
